Initial public commit of golem
Golem allows you to run many things based on commits to a git
repository. Think of it as a hooks/post-receive on steroids. Or git
meets continuous integration.

One example project is already included: daily deb and rpm builds from
the git.git next branch.

Documentation is of course missing :-)